WIP refactor model
