+++ allel/model/__init__.py
@@ -3,19 +3,23 @@
 from __future__ import absolute_import, print_function, division
 
 
-from allel.model import ndarray
-from allel.model import chunked
+from . import ndarray
+from .ndarray import *
+from . import util
+from .util import create_allele_mapping, locate_fixed_differences, locate_private_alleles
+
+# from allel.model import chunked
 
 # experimental
-try:
-    import dask.array as _da
-    from allel.model import dask
-except ImportError:
-    pass
+# try:
+#     import dask.array as _da
+#     from allel.model import dask
+# except ImportError:
+#     pass
 
 # deprecated
-try:
-    import bcolz as _bcolz
-    from allel.model import bcolz
-except ImportError:
-    pass
+# try:
+#     import bcolz as _bcolz
+#     from allel.model import bcolz
+# except ImportError:
+#     pass
