+++ allel/model/ndarray/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, print_function, division
+
+
+from .arrays import GenotypeArray, GenotypeVector, HaplotypeArray, AlleleCountsArray
+from .indexes import SortedIndex, UniqueIndex, SortedMultiIndex
+from .tables import VariantTable, FeatureTable
