+++ allel/test/model/test_api.py
