+++ allel/test/model/test_bcolz.py
