+++ allel/test/model/test_chunked.py
