+++ requirements.txt
@@ -9,6 +9,7 @@ petl
 seaborn
 toolz
 dask
+ipython
 h5py --no-binary :all:
 bcolz --no-binary :all:
 zarr --no-binary :all:
